+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/AudioVideoFacade.swift
@@ -9,4 +9,8 @@
 import Foundation
 
 @objc public protocol AudioVideoFacade: AudioVideoControllerFacade, RealtimeControllerFacade,
-    DeviceController, VideoTileControllerFacade, ActiveSpeakerDetectorFacade, ContentShareController {}
+    DeviceController,
+    VideoTileControllerFacade,
+    ActiveSpeakerDetectorFacade,
+    ContentShareController,
+    EventAnalyticsFacade {}